More Collectibles
Made feathers a collectible along with the carrots. Implemented a ring around the minimap as the stamina wheel. Hooked up UI for the collectibles. Re-sized the colliders to keep the player on the map with the new speed. Player reduced to quarter speed if they haven't picked up a carrot in the past minute. Adjusted hitboxes of turning buttons.