Basic Wolf AI Completed
the base functionality for the wolf AI has been completed. I fixed wandering to not give a change in direction every frame, tested to make sure seeking was working, created methods to help with obstacle avoidance. While avoidance isn't 100% of the time it does noticeably help with the wolves' pathing. Added a buffer zone to the invisible walls to steer the wolves back towards the center of the playspace if they wander too far to the edge so that they don't get stuck there and glitch out.