Carrot model, wolf speed, and turning
I exported the carrot model from Maya as a DAE file, and it's in the models folder, and the prefab is in the prefabs folder.

I added a timer for wolf speed marginally increasing every 10 seconds so that the game gets harder the longer the run goes on, but we might scrap this idea (playtesting might be required).  There's also a timer for the player in the GameManager script that does something similar but for bunny stamina, but that's commented out for now.

I also changed the quick turning (tap turning) angle to 30 degrees instead of 45 because 45 is really difficult to playtest with when I'm trying to collect things.

One last small thing, the low-poly nature models spawn at a certain height and then drop down, which looks unnatural, so I tried decreasing their spawn height by dividing the height offset by 4.0 instead of 2.0 in the SpawnObstacles() method, but it doesn't seem to make that much of a difference.

Also removed the extra box collider from the bunny so that collections aren't counted twice for the same collectable.