Sound Effects, Bugfixes, UI Aesthetics, and More!
Replaced sound effects in scene with free ones and credited where appropriate. Fixed some more sound player bugs. Created a range indicator on the player that speeds up the music if a wolf is near and chasing the player. Created a red UI border that flashes at the player during this scenario as well.