Obstacle Loading/Unloading
Started disabling the mesh renderers of objects if they are out of a certain range of the player so that it doesn't render all objects in the scene at all times. Drastically increases framerate.
Also disabled unity water, as it ate 20~60 frames depending on where on the map you were looking out towards the water. Need to look into a fog to hide the obstacle loading at the edge of the player's new sight radius.